+++ CHANGELOG.md
@@ -4,7 +4,7 @@ v1.5.1
 Added:
 
 * Added a "Start on Tyre from Fastest Lap" option to Race Weekend Filtering. You can use this to force an entrant for a session to start on the tyre they used to complete their fastest lap in the previous session. This might be useful when simulating F1-style qualifications.
-
+* Added configurable negative points modifiers for championships for crashes and cuts.
 
 Fixes: